+++ kibitzr/transformer/__init__.py
@@ -1,0 +1,1 @@
+from .factory import transform_factory  # noqa