+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+## Removed
+- cut and sort transforms (superseeded by bash)
 
 ## [3.0.11] - 2017-04-30
 ## Added